(2023.6, model) 2023.6 maven compilation is correct now on osx. Simplifying poms of children projects. Fixes on long for sizes on BoundedBuffer Report, transforming the only xtend class to a java class.

+++ turnus.model/src/turnus/model/analysis/buffers/BoundedBuffersReport.java
@@ -189,7 +189,7 @@ public interface BoundedBuffersReport extends AnalysisReport {
 	 * @model default="0" transient="true" changeable="false" volatile="true" derived="true"
 	 * @generated
 	 */
-	int getBitSize();
+	long getBitSize();
 
 	/**
 	 * Return the buffer size configuration contained in this report
